MaxScore.js
/problem: find maximum score given an array of numbers
/*it takes an array of number as input and output an int thats the biggest value added

catch: if num[i] is selected, and num[i-1] and/or num[i+1] has the same value as num[i],
all neighboring numbers that has a neighbor of num[i] will be deleted.

If any element is chosen to be removed and added to the sum,
then any other element equal to one of the neighbors will be removed without being added to the score.
In the first case, by removing 2 (index 1) it has a neighbor of 2, hence all 2s will be removed in one round.

for example, if we have